pyproject.toml
[tool.poetry]
name = "binance-portfolio-tracker"
version = "0.1.0"
description = "Webapp for binance portfolio tracking and visualization"
authors = ["Jeevan Jamakayala <jeevan.jamakayala@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
dash = "^2.5.1"
python-binance = "^1.0.16"
dash-bootstrap-components = "^1.1.0"
pandas = "^1.4.2"
python-dotenv = "^0.20.0"
SQLAlchemy = "^1.4.37"
dateparser = "^1.1.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
